requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
google-genai>=0.5.0
langgraph>=0.1.0
redis[asyncio]>=5.0.0
chromadb>=0.5.0
pypdf>=4.0.0
python-multipart>=0.0.9
